compiling in windows; building lexer library docs

+++ 4ed_app_models.h
@@ -80,7 +80,7 @@ struct Models{
     struct Live_Views *live_set;
     Editing_File *message_buffer;
     Editing_File *scratch_buffer;
-    f32 default_display_width;
+    i32 default_display_width;
     
     char hot_dir_base_[256];
     Hot_Directory hot_directory;
